Modification inscription , problème lors de l'inscription d'un membre

+++ ProjetCycliste/src/be/Denis/Model/Membre.java
@@ -1,6 +1,6 @@
 package be.Denis.Model;
 
-import java.util.Date;
+import java.sql.Date;
 import java.sql.SQLException;
 
 import be.Denis.DAO.AbstractDAOFactory;
@@ -11,7 +11,7 @@ import be.Denis.Vue.VueMembre;
 public class Membre extends Personne {
 	
 	
-	public Membre(String nom, String prenom, Date dateNaiss, String sexe, String categorie,String adresse, int numeroMaison, int codePostal, String ville, int numTel, String eMail, String login, String password, String fonction){
+	public Membre(String nom, String prenom, Date dateNaiss, String sexe, String categorie,String adresse, int numeroMaison, int codePostal, String ville, long numTel, String eMail, String login, String password, String fonction){
 		super(nom,prenom,dateNaiss, sexe,categorie,adresse,numeroMaison,codePostal,ville,numTel,eMail,login,password,fonction);
 	}
